LayoutTests/imported/w3c/web-platform-tests/XMLHttpRequest/send-content-type-charset-expected.txt


FAIL header with invalid MIME type is not changed assert_equals: expected "text; charset=ascii" but got "text; charset=UTF-8"
PASS known charset but bogus header - missing MIME type 
PASS bogus charset and bogus header - missing MIME type 
FAIL Correct text/plain MIME with charset assert_equals: expected "text/plain;charset=utf-8" but got "text/plain;charset=UTF-8"
PASS If no charset= param is given, implementation should not add one - unknown MIME 
PASS If no charset= param is given, implementation should not add one - known MIME 
FAIL charset given but wrong, fix it (unknown MIME, bogus charset) assert_equals: expected "text/x-thepiano;charset=UTF-8" but got "text/x-thepiano;charset= UTF-8"
FAIL charset given but wrong, fix it (known MIME, bogus charset) assert_equals: expected "text/plain;charset=utf-8;charset=UTF-8" but got "text/plain;charset=UTF-8;charset=UTF-8"
PASS charset given but wrong, fix it (known MIME, actual charset) 
PASS If multiple charset parameters are given, all should be rewritten 
PASS No content type set, give MIME and charset 
